fix ten bugs across merge, trimse, and stats; bump to v1.2.1
trimse: reads without a detectable adapter are now kept rather than
silently discarded (trim_read returned 0 for both "not found" and
"found at pos 0", causing all adapter-free reads to be dropped)

trimse: minLen now defaults to DEFAULT_MIN_LEN (15) as advertised in
the help text, rather than 0

trimse/merge: trim3p quality threshold comparison is now done in double
precision — the previous (int) cast before the multiply snapped every
computed Phred to a multiple of 10, breaking the threshold for any
trimQ that is not a multiple of 10

stats: fclose(NULL) crash fixed by guarding all optional file pointer
closes; stats now default to stderr as the help text states, instead of
reopening /dev/stdout; -z error message corrected from -o to -O

stats: statMedian now accumulates read counts (not base-weighted sums)
and uses ceiling division, giving the correct lower median read length

stats: gcHist and ppCountHist->d were never freed (memory leaks)

bfqutils: typo "single-rend" -> "single-end" in help text

+++ src/bfqmerge.c
@@ -245,7 +245,7 @@ static inline int trim3p(kseq_t *read, const int trimQ, const int trimW) {
                 avgQual += pow(10, ((double) (read->qual.s[j] - 33) / (-10.0)));
             }
             avgQual /= trimW;
-            if (((int) log10(avgQual) * -10.0) < trimQ) {
+            if ((log10(avgQual) * -10.0) < (double) trimQ) {
                 toTrim = i;
             } else {
                 break;
